+++ test/tcid_checker_test.dart
@@ -3,28 +3,35 @@ import 'package:test/test.dart';
 
 void main() {
   group('A group of tests', () {
+    test('Skip real Citizen ID', () {
+      expect(controlID("02345678982", false, false), isFalse);
+      expect(controlID("02345678982", true, false), isTrue);
+    });
     test('All of False', () {
-      expect(controlID(11111111111), isFalse);
-      expect(controlID(12345678912), isFalse);
+      expect(controlID("11111111111", false, true), isFalse);
+      expect(controlID("12345678912", false, true), isFalse);
     });
 
     test('Less than 11 digits', () {
-      expect(controlID(111111111), isFalse);
-      expect(controlID(123456789), isFalse);
+      expect(controlID("111111111", false, true), isFalse);
+      expect(controlID("123456789", false, true), isFalse);
     });
 
     test('More than 11 digits', () {
-      expect(controlID(11111111111111), isFalse);
-      expect(controlID(123456789122222), isFalse);
+      expect(controlID("11111111111111", false, true), isFalse);
+      expect(controlID("123456789122222", false, true), isFalse);
     });
 
     test('Validate', () async {
-      expect(await validateID(11111111111, "ali", "veli", 1991), isFalse);
-      expect(await validateForeignID(11111111111, "jack", "delay", 1, 1, 1900),
+      expect(
+          await validateID("11111111111", "ali", "veli", 1991, false), isFalse);
+      expect(
+          await validateForeignID(
+              "11111111111", "jack", "delay", 1, 1, 1900, false),
           isFalse);
       expect(
-          await validatePersonAndCard(11111111111, "ali", "veli", false, 1,
-              false, 1, false, 1900, 'a15', 796544, 'y02n45764'),
+          await validatePersonAndCard("11111111111", "ali", "veli", false, 1,
+              false, 1, false, 1900, 'a15', 796544, 'y02n45764', false),
           isFalse);
     });
   });